refactor: moved the interfaces in new file
