+++ src/store/index.ts
@@ -4,8 +4,6 @@ import chatReducer from './chatsSlice';
 import editReducer from './editMessageSlice';
 import filterReducer from './filtersSlice';
 
-export type IRootState = ReturnType<typeof store.getState>;
-
 export const store = configureStore({
   reducer: {
     auth: authReducer,
